# api/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "2.0.0",
        "title": "Openmoney API",
        "description": "Openmoney API",
        "termsOfService": "http://openmoney.cc/index.php?page=terms",
        "contact": {
            "name": "Openmoney Development Team",
            "url": "http://openmoney.cc"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0"
        }
    },
    "host": "deefactorial.proudleo.com",
    "basePath": "/V2",
    "schemes": [
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json",
        "text/html"
    ],
    "paths": {
...
        "/stewards/{stewardname}/login": {
            "get": {
                "x-swagger-router-controller": "Auth",
                "operationId": "loginGet",
                "parameters": [
                    {
                        "name": "stewardname",
                        "in": "path",
                        "type": "string",
                        "required": true,
                        "maxLength": 255,
                        "pattern": "^[A-Za-z0-9_.-]+$"
                    }
                ],
                "tags": [
                    "auth"
                ],
                "produces": [
                    "text/plain"
                ],
                "responses": {
                    "200": {
                        "description": "A login page for the steward",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "x-swagger-router-controller": "Auth",
                "operationId": "loginPost",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "type": "string",
                        "pattern": "^Basic [A-Za-z0-9/+=.]+$"
                    },
                    {
                        "name": "stewardname",
                        "in": "path",
                        "type": "string",
                        "required": true,
                        "maxLength": 255,
                        "pattern": "^[A-Za-z0-9_.-]+$"
                    }
                ],
                "tags": [
                    "auth"
                ],
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "text/plain"
                ],
                "responses": {
                    "302": {
                        "description": "Redirect to account page",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/stewards/{stewardname}/account": {
            "get": {
                "x-swagger-router-controller": "Auth",
                "operationId": "accountGet",
                "parameters": [
                    {
                        "name": "stewardname",
                        "in": "path",
                        "type": "string",
                        "required": true,
                        "maxLength": 255,
                        "pattern": "^[A-Za-z0-9_.-]+$"
                    }
                ],
                "tags": [
                    "auth"
                ],
                "produces": [
                    "text/html"
                ],
                "responses": {
                    "200": {
                        "description": "An account page",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/stewards/{stewardname}/oauth/token": {
            "post": {
                "x-swagger-router-controller": "Auth",
                "operationId": "oauthAccessTokenPost",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "type": "string",
                        "pattern": "^Bearer|Basic [A-Za-z0-9/+=.]+$"
                    },
                    {
                        "name": "stewardname",
                        "in": "path",
                        "type": "string",
                        "required": true,
                        "maxLength": 255,
                        "pattern": "^[A-Za-z0-9_.-]+$"
                    },
                    {
                        "name": "access_token_request",
                        "in": "body",
                        "description": "Access Token Request Object",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/access_token_request"
                        }
                    }
                ],
                "description": "Gets an authorization token",
                "tags": [
                    "auth"
                ],
                "summary": "Gets an authorization token",
                "security": [
                    {
                        "password": []
                    },
                    {
                        "oauth2Refresh": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Access Token is returned",
                        "schema": {
                            "$ref": "#/definitions/access_token_response"
                        }
                    },
                    "default": {
                        "description": "Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorModel"
                        }
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "oauth2Password": {
            "type": "oauth2",
            "scopes": {
                "read:stewards": "Read access to stewards",
                "write:stewards": "Write access to stewards",
                "read:accounts": "Read access to accounts",
                "write:accounts": "Write access to accounts",
                "read:currencies": "Read access to currencies",
                "write:currencies": "Write access to currencies",
                "manage:currencies": "Manage currencies",
                "read:namespaces": "Read access to namespaces",
                "write:namespaces": "Write access to namespaces",
                "manage:namespaces": "Manage namespaces",
                "read:journals": "Read access to journals",
                "write:journals": "Write access  to journals"
            },
            "flow": "password",
            "tokenUrl": "https://deefactorial.proudleo.com/V2/stewards/{stewardname}/token"
        },
        "oauth2Implicit": {
            "type": "oauth2",
            "scopes": {
                "read:stewards": "Read access to stewards",
                "write:stewards": "Write access to stewards",
                "read:accounts": "Read access to accounts",
                "write:accounts": "Write access to accounts",
                "read:currencies": "Read access to currencies",
                "write:currencies": "Write access to currencies",
                "manage:currencies": "Manage currencies",
                "read:namespaces": "Read access to namespaces",
                "write:namespaces": "Write access to namespaces",
                "manage:namespaces": "Manage namespaces",
                "read:journals": "Read access to journals",
                "write:journals": "Write access  to journals"
            },
            "flow": "implicit",
            "authorizationUrl": "http://deefactorial.proudleo.com/V2/steward/{stewardname}/oauth/dialoge"
        },
        "oauth2ClientCredentials": {
            "type": "oauth2",
...
    "definitions": {
...
        "access_token_request": {
            "type": "object",
            "minProperties": 2,
            "maxProperties": 3,
            "required": [
                "grant_type"
            ],
            "properties": {
                "grant_type": {
                    "type": "string",
                    "maxLength": 20,
                    "pattern": "^[A-Za-z_]+$",
                    "description": "one of code, password, refresh_token, client_credentials"
                },
                "client_id": {
                    "type": "string",
                    "minLength": 5,
                    "maxLength": 800,
                    "pattern": "^[A-Za-z0-9\\\\\\s/=+-]*$",
                    "description": "Stewards Public Key"
                },
                "client_secret": {
                    "type": "string",
                    "minLength": 32,
                    "maxLength": 1600,
                    "pattern": "^[A-Za-z0-9\\\\\\\\\\\\s/=+-]*$",
                    "description": "Client Secret"
                },
                "username": {
                    "type": "string",
                    "pattern": "^[A-Za-z0-9_.-]+$"
                },
                "password": {
                    "type": "string"
                },
                "code": {
                    "type": "string",
                    "minLength": 39,
                    "maxLength": 1600,
                    "pattern": "^[A-Za-z0-9]+$",
                    "description": "authorization code receieved during /oauth/authorize response"
                }
            }
        },
        "access_token_response": {
            "type": "object",
            "minProperties": 5,
            "maxProperties": 6,
            "required": [
                "access_token",
                "refresh_token",
                "expires",
                "scope",
                "token_type"
            ],
            "properties": {
                "access_token": {
                    "type": "string",
                    "minLength": 32,
                    "pattern": "^[A-Za-z0-9=/+]+$",
                    "description": "This is the access token used to authenticate successive requests."
                },
                "refresh_token": {
                    "type": "string",
                    "minLength": 32,
                    "pattern": "^[A-Za-z0-9=/+]+$",
                    "description": "This is the refresh token used to refresh access tokens."
                },
                "token_type": {
                    "type": "string",
                    "pattern": "^Bearer$"
                },
                "expires": {
                    "type": "string",
                    "description": "time when token expires"
                },
                "scope": {
                    "type": "array",
                    "items": {
                        "type": "string",
                        "minLength": 5,
                        "maxLength": 32
                    }
                }
            }
        },
...
        "ErrorModel": {
            "type": "object",
            "required": [
                "code",
                "message"
            ],
            "properties": {
                "code": {
                    "type": "integer",
                    "format": "int32"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}